COMP: disable libpsl for curl
LibPSL in curl is intended to deal with cookies, preventing
usage of supercookies (nope that is not a superhero).
As we don't use cookies with curl, we don't need it. Plus libpsl is
difficult to install with Windows.
For more information see this note https://github.com/curl/curl/issues/16486#issuecomment-2684123924